+++ test/Stub/MiddlewareFactoryStub.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sirix\Mezzio\Routing\Contracts\Test\Stub;
+
+use Psr\Container\ContainerInterface;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
+use Psr\Http\Server\RequestHandlerInterface;
+use Sirix\Mezzio\Routing\Contracts\MiddlewareFactoryInterface;
+use Sirix\Mezzio\Routing\Contracts\MiddlewareSpecification;
+
+final readonly class MiddlewareFactoryStub implements MiddlewareFactoryInterface
+{
+    public function create(ContainerInterface $container, MiddlewareSpecification $specification): MiddlewareInterface
+    {
+        return new class implements MiddlewareInterface {
+            public function process(ServerRequestInterface $request, RequestHandlerInterface $handler): ResponseInterface
+            {
+                return $handler->handle($request);
+            }
+        };
+    }
+}